requirements.txt
asgiref==3.7.2
astroid==2.15.6
autopep8==2.0.2
colorama==0.4.6
dill==0.3.6
Django==4.2.3
django-ckeditor==6.6.1
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory_boy==3.3.1
Faker==33.0.0
flake8==6.1.0
gunicorn==20.1.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==10.0.0
platformdirs==3.8.1
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.7.0
pylint==2.17.4
python-dateutil==2.9.0.post0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
typing_extensions==4.12.2
tzdata==2023.3
whitenoise==6.5.0
wrapt==1.15.0
